Revert "Add cmake"
