--- deps/readme
@@ -1,1 +1,0 @@
-This is a placeholder for the pre complied dependencies
