+++ emulation/emulation.vcproj
@@ -1,0 +1,315 @@
+<?xml version="1.0" encoding="windows-1251"?>
+<VisualStudioProject
+	ProjectType="Visual C++"
+	Version="9.00"
+	Name="emulation"
+	ProjectGUID="{28FFDB5C-3CB0-4166-9E36-81A54CCFB1BC}"
+	RootNamespace="emulation"
+	Keyword="Win32Proj"
+	TargetFrameworkVersion="196613"
+	>
+	<Platforms>
+		<Platform
+			Name="Win32"
+		/>
+	</Platforms>
+	<ToolFiles>
+	</ToolFiles>
+	<Configurations>
+		<Configuration
+			Name="Debug|Win32"
+			OutputDirectory="$(SolutionDir)$(ConfigurationName)"
+			IntermediateDirectory="$(ConfigurationName)"
+			ConfigurationType="1"
+			CharacterSet="1"
+			>
+			<Tool
+				Name="VCPreBuildEventTool"
+			/>
+			<Tool
+				Name="VCCustomBuildTool"
+			/>
+			<Tool
+				Name="VCXMLDataGeneratorTool"
+			/>
+			<Tool
+				Name="VCWebServiceProxyGeneratorTool"
+			/>
+			<Tool
+				Name="VCMIDLTool"
+			/>
+			<Tool
+				Name="VCCLCompilerTool"
+				Optimization="0"
+				AdditionalIncludeDirectories="&quot;C:\Program Files\HDF_Group\HDF5\1.8.16\include&quot;;&quot;C:\Users\Anton\Documents\Visual Studio 2008\Projects\CPML_TFSF_DISPERSION\pugixml-1.7\src&quot;;&quot;C:\Program Files\VTK\include\vtk-6.3&quot;"
+				PreprocessorDefinitions="WIN32;_DEBUG;_CONSOLE"
+				MinimalRebuild="true"
+				BasicRuntimeChecks="3"
+				RuntimeLibrary="3"
+				OpenMP="true"
+				UsePrecompiledHeader="0"
+				WarningLevel="3"
+				DebugInformationFormat="4"
+			/>
+			<Tool
+				Name="VCManagedResourceCompilerTool"
+			/>
+			<Tool
+				Name="VCResourceCompilerTool"
+			/>
+			<Tool
+				Name="VCPreLinkEventTool"
+			/>
+			<Tool
+				Name="VCLinkerTool"
+				AdditionalDependencies="libhdf5.lib pugixmls.lib vtkCommonCore-6.3.lib vtkCommonDataModel-6.3.lib vtkIOXML-6.3.lib"
+				LinkIncremental="2"
+				AdditionalLibraryDirectories="&quot;C:\Program Files\HDF_Group\HDF5\1.8.16\lib&quot;;&quot;C:\Users\Anton\Documents\Visual Studio 2008\Projects\CPML_TFSF_DISPERSION\pugixml-1.7\scripts\vs2008\x32&quot;;&quot;C:\Program Files\VTK\lib&quot;"
+				GenerateDebugInformation="true"
+				SubSystem="1"
+				TargetMachine="1"
+			/>
+			<Tool
+				Name="VCALinkTool"
+			/>
+			<Tool
+				Name="VCManifestTool"
+			/>
+			<Tool
+				Name="VCXDCMakeTool"
+			/>
+			<Tool
+				Name="VCBscMakeTool"
+			/>
+			<Tool
+				Name="VCFxCopTool"
+			/>
+			<Tool
+				Name="VCAppVerifierTool"
+			/>
+			<Tool
+				Name="VCPostBuildEventTool"
+			/>
+		</Configuration>
+		<Configuration
+			Name="Release|Win32"
+			OutputDirectory="$(SolutionDir)$(ConfigurationName)"
+			IntermediateDirectory="$(ConfigurationName)"
+			ConfigurationType="1"
+			CharacterSet="1"
+			WholeProgramOptimization="1"
+			>
+			<Tool
+				Name="VCPreBuildEventTool"
+			/>
+			<Tool
+				Name="VCCustomBuildTool"
+			/>
+			<Tool
+				Name="VCXMLDataGeneratorTool"
+			/>
+			<Tool
+				Name="VCWebServiceProxyGeneratorTool"
+			/>
+			<Tool
+				Name="VCMIDLTool"
+			/>
+			<Tool
+				Name="VCCLCompilerTool"
+				Optimization="2"
+				EnableIntrinsicFunctions="true"
+				AdditionalIncludeDirectories="&quot;C:\Program Files\HDF_Group\HDF5\1.8.16\include&quot;;&quot;C:\Users\Anton\Documents\Visual Studio 2008\Projects\CPML_TFSF_DISPERSION\pugixml-1.7\src&quot;;&quot;C:\Program Files\VTK\include\vtk-6.3&quot;"
+				PreprocessorDefinitions="WIN32;NDEBUG;_CONSOLE"
+				RuntimeLibrary="2"
+				EnableFunctionLevelLinking="true"
+				OpenMP="true"
+				UsePrecompiledHeader="0"
+				WarningLevel="3"
+				DebugInformationFormat="3"
+			/>
+			<Tool
+				Name="VCManagedResourceCompilerTool"
+			/>
+			<Tool
+				Name="VCResourceCompilerTool"
+			/>
+			<Tool
+				Name="VCPreLinkEventTool"
+			/>
+			<Tool
+				Name="VCLinkerTool"
+				AdditionalDependencies="libhdf5.lib pugixmls.lib vtkCommonCore-6.3.lib vtkCommonDataModel-6.3.lib vtkIOXML-6.3.lib"
+				LinkIncremental="1"
+				AdditionalLibraryDirectories="&quot;C:\Program Files\HDF_Group\HDF5\1.8.16\lib&quot;;&quot;C:\Users\Anton\Documents\Visual Studio 2008\Projects\CPML_TFSF_DISPERSION\pugixml-1.7\scripts\vs2008\x32&quot;;&quot;C:\Program Files\VTK\lib&quot;"
+				GenerateDebugInformation="true"
+				SubSystem="1"
+				OptimizeReferences="2"
+				EnableCOMDATFolding="2"
+				TargetMachine="1"
+			/>
+			<Tool
+				Name="VCALinkTool"
+			/>
+			<Tool
+				Name="VCManifestTool"
+			/>
+			<Tool
+				Name="VCXDCMakeTool"
+			/>
+			<Tool
+				Name="VCBscMakeTool"
+			/>
+			<Tool
+				Name="VCFxCopTool"
+			/>
+			<Tool
+				Name="VCAppVerifierTool"
+			/>
+			<Tool
+				Name="VCPostBuildEventTool"
+			/>
+		</Configuration>
+	</Configurations>
+	<References>
+	</References>
+	<Files>
+		<Filter
+			Name="Source Files"
+			Filter="cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx"
+			UniqueIdentifier="{4FC737F1-C7A5-4376-A066-2A32D752A2FF}"
+			>
+			<Filter
+				Name="fdtd"
+				>
+				<File
+					RelativePath=".\src\fdtd\debye.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\fdtd\drude.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\fdtd\fdtd.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\fdtd\fdtd_bnd.c"
+					>
+					<FileConfiguration
+						Name="Debug|Win32"
+						ExcludedFromBuild="true"
+						>
+						<Tool
+							Name="VCCLCompilerTool"
+						/>
+					</FileConfiguration>
+					<FileConfiguration
+						Name="Release|Win32"
+						ExcludedFromBuild="true"
+						>
+						<Tool
+							Name="VCCLCompilerTool"
+						/>
+					</FileConfiguration>
+				</File>
+				<File
+					RelativePath=".\src\fdtd\fdtd_cpml.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\fdtd\fdtd_nfff.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\fdtd\fdtd_tfsf.c"
+					>
+				</File>
+			</Filter>
+			<Filter
+				Name="exec"
+				>
+				<File
+					RelativePath=".\src\exec\calc_sig_cur.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\cells.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\globals.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\main.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\medium.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\reports.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\setpar.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\exec\time.c"
+					>
+				</File>
+			</Filter>
+			<Filter
+				Name="save"
+				>
+				<File
+					RelativePath=".\src\save\auto_save.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\save\save_dat.c"
+					>
+				</File>
+				<File
+					RelativePath=".\src\save\save_vtr.cpp"
+					>
+				</File>
+			</Filter>
+		</Filter>
+		<Filter
+			Name="Header Files"
+			Filter="h;hpp;hxx;hm;inl;inc;xsd"
+			UniqueIdentifier="{93995380-89BD-4b04-88EB-625FBE52EBFB}"
+			>
+			<File
+				RelativePath=".\include\constants.h"
+				>
+			</File>
+			<File
+				RelativePath=".\include\data.h"
+				>
+			</File>
+			<File
+				RelativePath=".\include\data_types.h"
+				>
+			</File>
+			<File
+				RelativePath=".\include\fdtd.h"
+				>
+			</File>
+			<File
+				RelativePath=".\include\functions.h"
+				>
+			</File>
+			<File
+				RelativePath=".\include\globals.h"
+				>
+			</File>
+		</Filter>
+	</Files>
+	<Globals>
+	</Globals>
+</VisualStudioProject>
