+++ emulation/include/constants.h
