+++ emulation/include/functions.h
