+++ emulation/include/globals.h
