+++ emulation/src/exec/medium.c
@@ -1,11 +1,13 @@
 ﻿
-/*! \file Calculate permeability, permittivity, conductivity and wavefront gradient
-* \author A.Kryukov anton.krv@gmail.com
+/*! \file 
+* \brief ‘ункции вычисл¤ющие усрдненную сеточную проводимость, диэлектрическую,
+         магнитную проницаемость, градиент фронта.
+* \author  рюков ј.ј. anton.krv@gmail.com
 */
 
 #include <math.h>
-#include "globals.h"
-#include "constants.h"
+#include "../../include/globals.h"
+#include "../../include/constants.h"
 
 /*!
 * ”средненна¤ сеточна¤ проводимость.
