+++ emulation/src/exec/reports.c
@@ -1,6 +1,6 @@
 ﻿
 /*! \file 
-* \brief Log functions.
+* \brief Функции вывода протокола работы программы.
 *
 * Данные функции используются для вывода протокола работы программы.
 * Протокол предназначен для диагностики процесса выполнения программы.
@@ -8,7 +8,7 @@
 * Все функции требующие обменов неявно синхронизируются.
 * Нет причины делать протокол, основанный на широковещательном опросе всех процессов.
 * Протокол по умолчанию выводится для нулевого.
-* \author A. Kryukov anton.krv@gmail.com
+* \author Крюков А.А. anton.krv@gmail.com
 */
 
 /*! Отменяет предупреждения об опастности функций вывода. */
@@ -17,7 +17,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "globals.h"
+#include "../../include/globals.h"
 
 /*! Поток вывода информации. */
 static FILE *_fp_log = NULL;
