+++ emulation/src/fdtd/debye.c
@@ -1,13 +1,13 @@
 ﻿
 /*! \file 
-* \brief Drude dispersion model.
-* \author А.Kryukov anton.krv@gmail.com
+* \brief функции расчета дисперсионной модели Дебая.
+* \author Крюков А.А. anton.krv@gmail.com
 */
 
-#include "globals.h"
-#include "constants.h"
-#include "functions.h"
-#include "fdtd.h"
+#include "../../include/globals.h"
+#include "../../include/constants.h"
+#include "../../include/functions.h"
+#include "../../include/fdtd.h"
 
 /*!
 * \brief Структура для описания параметров материала.
@@ -15,34 +15,33 @@
 typedef struct
 {
     int nPoles;//число полюсов
-    double *res_freq;//резонансная частота на данном полюсе
-    double *col_freq;//обратное время релаксации на данном полюсе
-    // время релаксации не может быть нулевым, а вот частота столкновений в плазме может
+    double *deltaEps;//изменение диэлектрической проницаемости на данном полюсе
+    double *tauEps;//время релаксации на данном полюсе
 
-    double *am;//вспомогательные массивы
+    double *km;//вспомогательные массивы
     double *bm;
-    double sig;//эффективная проводимость
-} _drudeMaterial;
+    double eps;//эффективная диэлектрическая проницаемость
+} _debyeMaterial;
 
-static _drudeMaterial matData;//материал
+static _debyeMaterial matData;//материал
 
 /*!
 * \brief Структура для хранения параметров расчета в точке.
 */
 typedef struct
 {
-    _drudeMaterial *material;//материал
+    _debyeMaterial *material;//материал
     double *Jpol;//компоненты поляризационного тока для каждого полюса
-} _drude;
+} _debye;
 
-_drude ****drude_Jx;/*!< массив x-компонеты поляризационного тока для электрического поля. */
-_drude ****drude_Jy;/*!< массив y-компонеты поляризационного тока для электрического поля. */
-_drude ****drude_Jz;/*!< массив z-компонеты поляризационного тока для электрического поля. */
+_debye ****debye_Jx;/*!< массив x-компонеты поляризационного тока для электрического поля. */
+_debye ****debye_Jy;/*!< массив y-компонеты поляризационного тока для электрического поля. */
+_debye ****debye_Jz;/*!< массив z-компонеты поляризационного тока для электрического поля. */
 
-static _drude *_init_data(_drudeMaterial *material)
+static _debye *_init_data(_debyeMaterial *material)
 {
-    _drude *data;
-    PARMS_malloc(data, 1, _drude);
+    _debye *data;
+    PARMS_malloc(data, 1, _debye);
 
     data->material = material;
     PARMS_malloc(data->Jpol, material->nPoles, double);
@@ -53,81 +52,81 @@ static _drude *_init_data(_drudeMaterial *material)
 }
 
 //инициализация дисперсионного материала
-void init_drude_data()
+void init_debye_data()
 {
     int i, j, k;
     grid _grd = &fdtd_grd;
 
     matData.nPoles = 1;
-    PARMS_malloc(matData.res_freq, matData.nPoles, double);
-    PARMS_malloc(matData.col_freq, matData.nPoles, double);
-    PARMS_malloc(matData.am, matData.nPoles, double);
+    PARMS_malloc(matData.deltaEps, matData.nPoles, double);
+    PARMS_malloc(matData.tauEps, matData.nPoles, double);
+    PARMS_malloc(matData.km, matData.nPoles, double);
     PARMS_malloc(matData.bm, matData.nPoles, double);
 
     for (i = 0; i < matData.nPoles; ++i) {
-        matData.res_freq[i] = 1.0e+09;
-        matData.col_freq[i] = 1.0e+02;
-        matData.am[i] = 0.0;
+        matData.deltaEps[i] = 1.0;
+        matData.tauEps[i] = 1.0/(_pi*5.0e08);
+        matData.km[i] = 0.0;
         matData.bm[i] = 0.0;
     }
 
-    matData.sig = 0.0;
+    matData.eps = 0.0;
 
-    PARMS_malloc(drude_Jx, _grd->nx, _drude***);
+    PARMS_malloc(debye_Jx, _grd->nx, _debye***);
     for (i = 0; i < _grd->nx; ++i) {
-        PARMS_malloc(drude_Jx[i], _grd->ny+1, _drude**);
+        PARMS_malloc(debye_Jx[i], _grd->ny+1, _debye**);
         for (j = 0; j < _grd->ny+1; ++j) {
-            PARMS_malloc(drude_Jx[i][j], _grd->nz+1, _drude*);
+            PARMS_malloc(debye_Jx[i][j], _grd->nz+1, _debye*);
             for (k = 0; k < _grd->nz+1; ++k) {
-                drude_Jx[i][j][k] = _init_data(&matData);
+                debye_Jx[i][j][k] = _init_data(&matData);
             }
         }
     }
 
-    PARMS_malloc(drude_Jy, _grd->ny, _drude***);
+    PARMS_malloc(debye_Jy, _grd->ny, _debye***);
     for (j = 0; j < _grd->ny; ++j) {
-        PARMS_malloc(drude_Jy[j], _grd->nz+1, _drude**);
+        PARMS_malloc(debye_Jy[j], _grd->nz+1, _debye**);
         for (k = 0; k < _grd->nz+1; ++k) {
-            PARMS_malloc(drude_Jy[j][k], _grd->nx+1, _drude*);
+            PARMS_malloc(debye_Jy[j][k], _grd->nx+1, _debye*);
             for (i = 0; i < _grd->nx+1; ++i) {
-                drude_Jy[j][k][i] = _init_data(&matData);
+                debye_Jy[j][k][i] = _init_data(&matData);
             }
         }
     }
 
-    PARMS_malloc(drude_Jz, _grd->nz, _drude***);
+    PARMS_malloc(debye_Jz, _grd->nz, _debye***);
     for (k = 0; k < _grd->nz; ++k) {
-        PARMS_malloc(drude_Jz[k], _grd->nx+1, _drude**);
+        PARMS_malloc(debye_Jz[k], _grd->nx+1, _debye**);
         for (i = 0; i < _grd->nx+1; ++i) {
-            PARMS_malloc(drude_Jz[k][i], _grd->ny+1, _drude*);
+            PARMS_malloc(debye_Jz[k][i], _grd->ny+1, _debye*);
             for (j = 0; j < _grd->ny+1; ++j) {
-                drude_Jz[k][i][j] = _init_data(&matData);
+                debye_Jz[k][i][j] = _init_data(&matData);
             }
         }
     }
 }
 
 //расчет дисперсионного материала на каждом шаге
-void update_drude_data(double ht)
+void update_debye_data(double ht)
 {
     int i;
 
-    matData.sig = 0.0;
+    matData.eps = 0.0;
 
     for (i = 0; i < matData.nPoles; ++i) {
-        matData.am[i] = (2.0 - ht*matData.col_freq[i])/(2.0 + ht*matData.col_freq[i]);
-        matData.bm[i] = _pi4*((matData.res_freq[i]*matData.res_freq[i]*ht)/(2.0 + ht*matData.col_freq[i]));
+        matData.km[i] = (2.0*matData.tauEps[i] - ht)/(2.0*matData.tauEps[i] + ht);
+        matData.bm[i] = _pi4*((2.0*matData.deltaEps[i]*ht)/(2.0*matData.tauEps[i] + ht));
 
-        matData.sig += matData.bm[i];
+        matData.eps += matData.bm[i];
     }
 
-    matData.sig *= 0.5;//получили поправку для проводимости
+    matData.eps *= 0.5;//получили поправку для проницаемости
 }
 
 //никогда не вызывать эти функции в CPML слое!!!!!
 
 //вычисление Ex в точке
-void calc_Ex_point_drude(int i, int j, int k)
+void calc_Ex_point_debye(int i, int j, int k)
 {
     double ht = fdtd_ht;
     grid _grd = &fdtd_grd;
@@ -136,36 +135,36 @@ void calc_Ex_point_drude(int i, int j, int k)
     double eps_tmp = fun_eps(i, j, k, EX_ID);// считаем, что диэлетрическая проницаемость для бесконечной частоты задана базовым материалом
     int l;
     double polarJx = 0.0;
-    double Ex;
-    _drudeMaterial *_mat = drude_Jx[i-1][j][k]->material;
+    double dEx;
+    _debyeMaterial *_mat = debye_Jx[i-1][j][k]->material;
 
     //вычисляем поляризационный ток
     for (l = 0; l < _mat->nPoles; ++l) {
-        polarJx += drude_Jx[i-1][j][k]->Jpol[l]*(1.0 + _mat->am[l]);
+        polarJx += debye_Jx[i-1][j][k]->Jpol[l]*(1.0 + _mat->km[l]);
     }
 
-    polarJx *= 0.5;
+    polarJx += 0.5;
 
-    sig_tmp += _mat->sig;//добавили дисперсионную поправку
+    eps_tmp += _mat->eps;//добавили дисперсионную поправку
 
     dhy = (fdtd_Hy[j][k][i-1] - fdtd_Hy[j][k-1][i-1])/_grd->dzi[k];
     dhz = (fdtd_Hz[k][i-1][j] - fdtd_Hz[k][i-1][j-1])/_grd->dyi[j];
 
-    Ex = fdtd_Ex[i-1][j][k];//Ex[n]
+    dEx = -fdtd_Ex[i-1][j][k];//-Ex[n]
 
     fdtd_Ex[i-1][j][k] = ((eps_tmp - 0.5*ht*sig_tmp)*fdtd_Ex[i-1][j][k]
                     + ht*_cl*(dhz - dhy) - ht*(_pi4*fdtd_Jx[i-1][j][k] + polarJx))/(eps_tmp + 0.5*ht*sig_tmp);
 
-    Ex += fdtd_Ex[i-1][j][k];//Ex[n+1] + Ex[n]
-    Ex *= 0.5;//(Ex[n+1] + Ex[n])/2
+    dEx += fdtd_Ex[i-1][j][k];//Ex[n+1] - Ex[n]
+    dEx /= ht;//(Ex[n+1] - Ex[n])/ht
 
     for (l = 0; l < _mat->nPoles; ++l) {//расчет поляризационного тока на след шаге
-        drude_Jx[i-1][j][k]->Jpol[l] = _mat->am[l]*drude_Jx[i-1][j][k]->Jpol[l] + _mat->bm[l]*Ex;
+        debye_Jx[i-1][j][k]->Jpol[l] = _mat->km[l]*debye_Jx[i-1][j][k]->Jpol[l] + _mat->bm[l]*dEx;
     }
 }
 
 //вычисление Ey в точке
-void calc_Ey_point_drude(int i, int j, int k)
+void calc_Ey_point_debye(int i, int j, int k)
 {
     double ht = fdtd_ht;
     grid _grd = &fdtd_grd;
@@ -174,36 +173,36 @@ void calc_Ey_point_drude(int i, int j, int k)
     double eps_tmp = fun_eps(i, j, k, EY_ID);// считаем, что диэлетрическая проницаемость для бесконечной частоты задана базовым материалом
     int l;
     double polarJy = 0.0;
-    double Ey;
-    _drudeMaterial *_mat = drude_Jy[j-1][k][i]->material;
+    double dEy;
+    _debyeMaterial *_mat = debye_Jy[j-1][k][i]->material;
 
     //вычисляем поляризационный ток
     for (l = 0; l < _mat->nPoles; ++l) {
-        polarJy += drude_Jy[j-1][k][i]->Jpol[l]*(1.0 + _mat->am[l]);
+        polarJy += debye_Jy[j-1][k][i]->Jpol[l]*(1.0 + _mat->km[l]);
     }
 
-    polarJy *= 0.5;
+    polarJy += 0.5;
 
-    sig_tmp += _mat->sig;//добавили дисперсионную поправку
+    eps_tmp += _mat->eps;//добавили дисперсионную поправку
 
     dhx = (fdtd_Hx[i][j-1][k] - fdtd_Hx[i][j-1][k-1])/_grd->dzi[k];
     dhz = (fdtd_Hz[k][i][j-1] - fdtd_Hz[k][i-1][j-1])/_grd->dxi[i];
 
-    Ey = fdtd_Ey[j-1][k][i];//Ey[n]
+    dEy = -fdtd_Ey[j-1][k][i];//-Ey[n]
 
     fdtd_Ey[j-1][k][i] = ((eps_tmp - 0.5*ht*sig_tmp)*fdtd_Ey[j-1][k][i]
                     + ht*_cl*(dhx - dhz) - ht*(_pi4*fdtd_Jy[j-1][k][i] + polarJy))/(eps_tmp + 0.5*ht*sig_tmp);
 
-    Ey += fdtd_Ey[j-1][k][i];//Ey[n+1] + Ey[n]
-    Ey *= 0.5;//(Ey[n+1] + Ey[n])/2
+    dEy += fdtd_Ey[j-1][k][i];//Ey[n+1] - Ey[n]
+    dEy /= ht;//(Ey[n+1] - Ey[n])/ht
 
     for (l = 0; l < _mat->nPoles; ++l) {//расчет поляризационного тока на след шаге
-        drude_Jy[j-1][k][i]->Jpol[l] = _mat->am[l]*drude_Jy[j-1][k][i]->Jpol[l] + _mat->bm[l]*Ey;
+        debye_Jy[j-1][k][i]->Jpol[l] = _mat->km[l]*debye_Jy[j-1][k][i]->Jpol[l] + _mat->bm[l]*dEy;
     }
 }
 
 //вычисление Ez в точке
-void calc_Ez_point_drude(int i, int j, int k)
+void calc_Ez_point_debye(int i, int j, int k)
 {
     double ht = fdtd_ht;
     grid _grd = &fdtd_grd;
@@ -212,30 +211,30 @@ void calc_Ez_point_drude(int i, int j, int k)
     double eps_tmp = fun_eps(i, j, k, EZ_ID);// считаем, что диэлетрическая проницаемость для бесконечной частоты задана базовым материалом
     int l;
     double polarJz = 0.0;
-    double Ez;
-    _drudeMaterial *_mat = drude_Jz[k-1][i][j]->material;
+    double dEz;
+    _debyeMaterial *_mat = debye_Jz[k-1][i][j]->material;
 
     //вычисляем поляризационный ток
     for (l = 0; l < _mat->nPoles; ++l) {
-        polarJz += drude_Jz[k-1][i][j]->Jpol[l]*(1.0 + _mat->am[l]);
+        polarJz += debye_Jz[k-1][i][j]->Jpol[l]*(1.0 + _mat->km[l]);
     }
 
-    polarJz *= 0.5;
+    polarJz += 0.5;
 
-    sig_tmp += _mat->sig;//добавили дисперсионную поправку
+    eps_tmp += _mat->eps;//добавили дисперсионную поправку
 
     dhx = (fdtd_Hx[i][j][k-1] - fdtd_Hx[i][j-1][k-1])/_grd->dyi[j];
     dhy = (fdtd_Hy[j][k-1][i] - fdtd_Hy[j][k-1][i-1])/_grd->dxi[i];
 
-    Ez = fdtd_Ez[k-1][i][j];//Ez[n]
+    dEz = -fdtd_Ez[k-1][i][j];//-Ez[n]
 
     fdtd_Ez[k-1][i][j] = ((eps_tmp - 0.5*ht*sig_tmp)*fdtd_Ez[k-1][i][j]
                     + ht*_cl*(dhy - dhx) - ht*(_pi4*fdtd_Jz[k-1][i][j] + polarJz))/(eps_tmp + 0.5*ht*sig_tmp);
 
-    Ez += fdtd_Ez[k-1][i][j];//Ez[n+1] + Ez[n]
-    Ez *= 0.5;//(Ez[n+1] + Ez[n])/2
+    dEz += fdtd_Ez[k-1][i][j];//Ez[n+1] - Ez[n]
+    dEz /= ht;//(Ez[n+1] - Ez[n])/ht
 
     for (l = 0; l < _mat->nPoles; ++l) {
-        drude_Jz[k-1][i][j]->Jpol[l] = _mat->am[l]*drude_Jz[k-1][i][j]->Jpol[l] + _mat->bm[l]*Ez;
+        debye_Jz[k-1][i][j]->Jpol[l] = _mat->km[l]*debye_Jz[k-1][i][j]->Jpol[l] + _mat->bm[l]*dEz;
     }
 }
