+++ emulation/src/fdtd/fdtd.c
@@ -1,15 +1,15 @@
 ﻿
 /*! \file 
 * \brief Функции расчета полей.
-* \author A. Kryukov. anton.krv@gmail.com
+* \author Крюков А.А. anton.krv@gmail.com
 */
 
 #include <math.h>
 #include <omp.h>
-#include "globals.h"
-#include "constants.h"
-#include "functions.h"
-#include "fdtd.h"
+#include "../../include/globals.h"
+#include "../../include/constants.h"
+#include "../../include/functions.h"
+#include "../../include/fdtd.h"
 
 //вычисление Ex в точке
 static void calc_Ex_point(int i, int j, int k)
