+++ emulation/src/fdtd/fdtd_cpml.c
@@ -1,14 +1,14 @@
 ﻿
 /*! \file 
-* \brief Convolutional PML model .
-* \author A. Kryukov anton.krv@gmail.com
+* \brief Функции расчета полей с идеальным поглащающим слоем на границе.
+* \author Крюков А.А. anton.krv@gmail.com
 */
 
 #include <math.h>
-#include "globals.h"
-#include "constants.h"
-#include "functions.h"
-#include "fdtd.h"
+#include "../../include/globals.h"
+#include "../../include/constants.h"
+#include "../../include/functions.h"
+#include "../../include/fdtd.h"
 
 pml_double **phu_ex_y_min;/*!< массив y- компоненты свертки для поля Ex. */
 pml_double **phu_ex_y_max;/*!< массив y- компоненты свертки для поля Ex. */
