+++ emulation/src/save/auto_save.c
@@ -1,10 +1,10 @@
 ﻿
 
 /*! \file 
-* \brief Save functions.
+* \brief Функции созранения.
 *
 * Функции позволяют сохранить результаты расчета для последующего его возобновления.
-* \author A.Kryukov anton.krv@gmail.com
+* \author Крюков А.А. anton.krv@gmail.com
 */
 
 /*! Отменяет предупреждения об опастности функций вывода. */
