+++ emulation/src/save/save_dat.c
@@ -1,9 +1,9 @@
 ﻿
 /*! \file 
-* \brief Save functions.
+* \brief Функции для сохранения результатов.
 *
 * Функции открытия потока вывода результатов расчета, записи результатов и закрытия потока вывода.
-* \author A.Kryukov anton.krv@gmail.com
+* \author Крюков А.А. anton.krv@gmail.com
 */
 
 /*! Отменяет предупреждения об опастности функций вывода. */
