--- tools/readme
@@ -1,1 +1,0 @@
-All the fun is here
